Rewrite Mark's motivation.
Mark deriving motivation from his rivalry with Melissa should come in time, but not immediately. It is not implicit that all break ups end in a spiteful cycle of vengeance. Watching their develop will be more interesting anyway.

It is more relate-able for him to derive motivation from the promise of co-op, because the dilemma of "I have this great game none of my friends will play with me" is much more common than "I place the suffering of my ex lover as a higher priority than my own enjoyment."